+++ CHANGES.rst
@@ -9,6 +9,7 @@ Changelog
 - Change SwissSign Idp endpoint URL
 - Show SAML2 metadata file configuration option
 - Patch pySAML2 so that POST is default binding in metadata configuration
+- Use Name instead of FriendlyName for core/derived assertion attributes
 
 0.1b1 (2010-10-22)
 ------------------